requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
rich>=13.0.0
matplotlib>=3.7.0
apscheduler>=3.10.0
